feat: port validations.js tap test to node-test
